requirements.txt
flake8
pandas~=1.4.2
ghostscript~=0.7
holidays~=0.13
folium~=0.12.1.post1
requests~=2.31.0
pdfminer.six==20220319
more_itertools~=9.1.0

# Install pdf2gtfs in editable/develop mode.
-e .

# Local packages. These will need to be installed manually first.
custom-conf
